separate wav2vec2 embeddings from fluent-specific
